[#43] feat: 매너점수 평가 팝업 모달 생성
